requirements.txt
aiogram==2.25.1
APScheduler==3.10.1
mysql_connector_repackaged==0.3.1
Pillow==10.0.0
Pillow==9.5.0
Pillow==10.2.0
python-dotenv==1.0.1
web3==6.4.0
